Seperated the ModelViewer into its own folder (needs more refactoring)
Fixes #447
